notification_service/pyproject.toml
[tool.poetry]
name = "test-task-notification-service"
version = "0.1.0"
description = ""
authors = ["Ilmir <you@Ðvalitovilmir29@gmail.com>"]
readme = "README.md"
packages = [{include = "test_task_notification_service"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
requests = "^2.28.1"
celery = "^5.2.7"
redis = "^4.3.4"
eventlet = "^0.33.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
